chore: drop NON-_isolated tests, rename _isolated tests
