+++ core/service/src/client/tests/centralized/misc_tests.rs
@@ -1,18 +1,19 @@
-// DEPRECATED: Isolated equivalents in `misc_tests_isolated.rs`
-// - test_central_health_endpoint_availability → test_central_health_endpoint_availability_isolated
-// - test_central_close_after_drop → test_central_close_after_drop_isolated
-// - test_largecipher → test_largecipher_isolated
-// TODO: Remove after migration complete.
+//! Centralized misc tests.
+//!
+//! These tests use the consolidated testing module and run in isolated
+//! temporary directories with pre-generated cryptographic material.
 
-use crate::client::test_tools::{get_health_client, get_status};
 use crate::client::tests::common::TIME_TO_SLEEP_MS;
 use crate::consts::TEST_CENTRAL_KEY_ID;
-use crate::consts::TEST_PARAM;
 use crate::engine::centralized::central_kms::RealCentralizedKms;
-use crate::vault::storage::file::FileStorage;
+use crate::testing::prelude::*;
+use crate::testing::utils::{get_health_client, get_status};
 use kms_grpc::kms_service::v1::core_service_endpoint_server::CoreServiceEndpointServer;
-use serial_test::serial;
 use std::collections::HashMap;
+use tonic::server::NamedService;
+use tonic_health::pb::HealthCheckRequest;
+use tonic_health::pb::health_check_response::ServingStatus;
+
 cfg_if::cfg_if! {
     if #[cfg(feature = "slow_tests")] {
         use tfhe::FheTypes;
@@ -23,20 +24,23 @@ cfg_if::cfg_if! {
         use crate::dummy_domain;
         use crate::client::client_wasm::Client;
         use kms_grpc::kms::v1::Empty;
+        use crate::consts::DEFAULT_CENTRAL_KEY_ID;
+        use crate::engine::centralized::central_kms::tests::{
+            new_priv_ram_storage_from_existing_keys, new_pub_ram_storage_from_existing_keys,
+        };
     }
 }
-use tonic::server::NamedService;
-use tonic_health::pb::HealthCheckRequest;
-use tonic_health::pb::health_check_response::ServingStatus;
 
-/// Check that the centralized health service is serving as soons as boot is completed.
+/// Check that the centralized health service is serving as soon as boot is completed.
 #[tokio::test]
-#[serial]
-async fn test_central_health_endpoint_availability() {
-    let (kms_server, _kms_client, _internal_client) =
-        crate::client::test_tools::centralized_handles(&TEST_PARAM, None).await;
+async fn test_central_health_endpoint_availability() -> Result<()> {
+    let env = CentralizedTestEnv::builder()
+        .with_test_name("health_endpoint")
+        .build()
+        .await?;
+
     tokio::time::sleep(tokio::time::Duration::from_millis(TIME_TO_SLEEP_MS)).await;
-    let mut health_client = get_health_client(kms_server.service_port)
+    let mut health_client = get_health_client(env.server.service_port)
         .await
         .expect("Failed to get health client");
     let service_name = <CoreServiceEndpointServer<
@@ -57,15 +61,38 @@ async fn test_central_health_endpoint_availability() {
         ServingStatus::Serving as i32,
         "Service is not in SERVING status. Got status: {status}"
     );
+
+    Ok(())
 }
 
 /// Validate that dropping the server signal triggers the server to shut down
 #[tokio::test]
-#[serial]
-async fn test_central_close_after_drop() {
-    tokio::time::sleep(tokio::time::Duration::from_millis(TIME_TO_SLEEP_MS)).await;
-    let (kms_server, kms_client, mut internal_client) =
-        crate::client::test_tools::centralized_handles(&TEST_PARAM, None).await;
+async fn test_central_close_after_drop() -> Result<()> {
+    use crate::consts::TEST_PARAM;
+
+    let env = CentralizedTestEnv::builder()
+        .with_test_name("close_after_drop")
+        .build()
+        .await?;
+
+    // Create additional storage instances for internal client
+    let pub_storage = FileStorage::new(Some(env.material_dir.path()), StorageType::PUB, None)?;
+    let client_storage =
+        FileStorage::new(Some(env.material_dir.path()), StorageType::CLIENT, None)?;
+
+    // Create internal client with isolated material
+    let pub_storage_map = HashMap::from([(1, pub_storage)]);
+    let mut internal_client = crate::client::client_wasm::Client::new_client(
+        client_storage,
+        pub_storage_map,
+        &TEST_PARAM,
+        None,
+    )
+    .await?;
+
+    let kms_server = env.server;
+    let kms_client = env.client;
+
     let mut health_client = get_health_client(kms_server.service_port)
         .await
         .expect("Failed to get health client");
@@ -88,7 +115,7 @@ async fn test_central_close_after_drop() {
         "Service is not in SERVING status. Got status: {status}"
     );
     let client_map = HashMap::from([(1, kms_client)]);
-    // Keep the server occupied so it won't shut down immidiately after dropping the handle
+    // Keep the server occupied so it won't shut down immediately after dropping the handle
     let (tasks, req_id) = crate::client::tests::common::send_dec_reqs(
         3,
         &TEST_CENTRAL_KEY_ID,
@@ -96,15 +123,14 @@ async fn test_central_close_after_drop() {
         &client_map,
         &mut internal_client,
         &[None],
-        None,
+        Some(env.material_dir.path()), // use isolated material path
     )
     .await;
     // Drop server
     drop(kms_server);
     // Get status and validate that it is not serving
     let status = get_status(&mut health_client, service_name).await.unwrap();
-    // Threshold servers will start serving as soon as they boot
-    // WARNING there is a risk this check fails if the server is shut down before was can complete the status check
+    // WARNING there is a risk this check fails if the server is shut down before we can complete the status check
     assert_eq!(
         status,
         ServingStatus::NotServing as i32,
@@ -116,27 +142,22 @@ async fn test_central_close_after_drop() {
     // And wait for public decryption to also be done
     let dec_resp_tasks = crate::client::tests::common::get_pub_dec_resp(&req_id, &client_map).await;
     let dec_resp_res = dec_resp_tasks.join_all().await;
-    // TODO the response for the server that were not dropped should actually be ok since we only drop one <=t server
+    // In centralized mode, dropping the single server means all responses should fail
     assert!(dec_resp_res.iter().all(|res| res.is_err()));
     // Check the server is no longer there
     assert!(get_status(&mut health_client, service_name).await.is_err());
+
+    Ok(())
 }
 
-// Validate bug-fix to ensure that the server fails gracefully when the ciphertext is too large
+/// Validate bug-fix to ensure that the server fails gracefully when the ciphertext is too large.
+///
+/// This test uses RAM storage with custom rate limiter config (not the builder pattern)
+/// because it needs fine-grained control over the rate limiter settings.
 #[cfg(feature = "slow_tests")]
 #[tokio::test(flavor = "multi_thread")]
-#[serial]
-async fn test_largecipher() {
-    use crate::{
-        consts::DEFAULT_CENTRAL_KEY_ID,
-        consts::DEFAULT_EPOCH_ID,
-        engine::centralized::central_kms::tests::{
-            new_priv_ram_storage_from_existing_keys, new_pub_ram_storage_from_existing_keys,
-        },
-    };
-
+async fn test_largecipher() -> Result<()> {
     let keys = get_default_keys().await;
-    let epoch_id = *DEFAULT_EPOCH_ID;
     let rate_limiter_conf = RateLimiterConfig {
         bucket_size: 100,
         pub_decrypt: 1,
@@ -146,18 +167,23 @@ async fn test_largecipher() {
         keygen: 1,
         new_epoch: 1,
     };
-    tokio::time::sleep(tokio::time::Duration::from_millis(TIME_TO_SLEEP_MS)).await;
+
+    // Setup with RAM storage and custom rate limiter
     let (kms_server, mut kms_client) = crate::client::test_tools::setup_centralized(
         new_pub_ram_storage_from_existing_keys(&keys.pub_fhe_keys)
             .await
             .unwrap(),
-        new_priv_ram_storage_from_existing_keys(&keys.centralized_kms_keys, &epoch_id)
-            .await
-            .unwrap(),
+        new_priv_ram_storage_from_existing_keys(
+            &keys.centralized_kms_keys,
+            &crate::consts::DEFAULT_EPOCH_ID,
+        )
+        .await
+        .unwrap(),
         None,
         Some(rate_limiter_conf),
     )
     .await;
+
     let ct = Vec::from([1_u8; 100000]);
     let fhe_type = FheTypes::Uint32;
     let ct_format = kms_grpc::kms::v1::CiphertextFormat::default();
@@ -187,8 +213,8 @@ async fn test_largecipher() {
             typed_ciphertexts,
             &request_id,
             &DEFAULT_CENTRAL_KEY_ID,
-            None,
-            None,
+            None, // context_id
+            None, // epoch_id
             &[],
         )
         .unwrap();
@@ -209,14 +235,20 @@ async fn test_largecipher() {
             .await;
     }
     // Check that we get a server error instead of a server crash
-    assert_eq!(response.as_ref().unwrap_err().code(), tonic::Code::Internal);
     assert!(
-        response
-            .err()
-            .unwrap()
-            .message()
-            .contains("Failed on requestID")
+        response.is_err(),
+        "Expected error response for large ciphertext, got Ok"
+    );
+    let err = response.unwrap_err();
+    assert_eq!(err.code(), tonic::Code::Internal);
+    assert!(
+        err.message().contains("finished with an error")
+            || err.message().contains("Failed on requestID"),
+        "Expected error message to indicate failure, got: {}",
+        err.message()
     );
     tracing::info!("aborting");
     kms_server.assert_shutdown().await;
+
+    Ok(())
 }
