+++ core/service/src/client/tests/centralized/mod.rs
@@ -3,13 +3,8 @@ mod custodian_backup_tests;
 mod custodian_context_tests;
 mod key_gen_tests;
 mod misc_tests;
-#[cfg(any(test, feature = "testing"))]
-mod misc_tests_isolated;
 #[cfg(feature = "slow_tests")]
 mod nightly_tests;
 mod public_decryption_tests;
-#[cfg(feature = "slow_tests")]
 mod restore_from_backup_tests;
-#[cfg(any(test, feature = "testing"))]
-mod restore_from_backup_tests_isolated;
 mod user_decryption_tests;
