+++ core/service/src/client/tests/threshold/mod.rs
@@ -3,6 +3,7 @@ mod crs_gen_tests;
 mod custodian_backup_tests;
 mod custodian_context_tests;
 mod key_gen_tests;
+#[cfg(any(feature = "testing", feature = "insecure"))]
 mod misc_tests;
 mod mpc_context_tests;
 #[cfg(feature = "slow_tests")]
@@ -10,7 +11,7 @@ mod mpc_epoch_tests;
 #[cfg(feature = "slow_tests")]
 mod nightly_tests;
 mod public_decryption_tests;
-#[cfg(feature = "slow_tests")]
+#[cfg(any(feature = "testing", feature = "insecure"))]
 mod restore_from_backup_tests;
 mod user_decryption_tests;
 
@@ -29,8 +30,3 @@ mod user_decryption_tests;
 #[cfg(any(feature = "testing", feature = "insecure"))]
 mod key_gen_tests_isolated;
 
-#[cfg(any(feature = "testing", feature = "insecure"))]
-mod misc_tests_isolated;
-
-#[cfg(any(feature = "testing", feature = "insecure"))]
-mod restore_from_backup_tests_isolated;
