+++ core/service/src/testing/mod.rs
@@ -68,8 +68,8 @@ pub mod utils;
 /// For threshold-specific helpers, import directly:
 /// ```
 /// use crate::client::tests::threshold::common::{
-///     threshold_key_gen_isolated,
-///     threshold_key_gen_secure_isolated,
+///     threshold_insecure_key_gen,
+///     threshold_key_gen_secure,
 /// };
 /// ```
 pub mod prelude {
